\name{bedload}
\alias{bedload}

\title{
Load and sort a narrow peaks bed file into dataframe.
}
\description{
Read and sort a narrow peaks bed file into a data frame.
}
\usage{
bedload(path)
}

\arguments{
  \item{path}{
Path to the bed file to load.
}
}
\details{
A simple bed file loader.
}
\value{
A dataframe of bed file with column names assigned according to the bed format standard.
}
\references{

}
\author{
Chinedu A. Anene, PhD
}
\note{

}

\seealso{
gffload
}
\examples{
See manuscript

  }

