\name{gffload}
\alias{gffload}

\title{
Load gff, gtf, gff3 into GRanges object.
}
\description{
Read a gtf, gff, gtf3 into GRanges for processing.
}
\usage{
gffload(path)
}

\arguments{
  \item{path}{
Path to the gff file to load.
}
}
\details{
A simple gff to GRanges loader.
}
\value{
GRanges object.
}
\references{
}
\author{
Chinedu A. Anene, PhD
}
\note{
}

\seealso{
bedload
}
\examples{
  }

